problem5/x/rental/types/message_repay_rental.go
package types

import (
	errorsmod "cosmossdk.io/errors"
	sdk "github.com/cosmos/cosmos-sdk/types"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
)

var _ sdk.Msg = &MsgRepayRental{}

func NewMsgRepayRental(creator string, id uint64) *MsgRepayRental {
	return &MsgRepayRental{
		Creator: creator,
		Id:      id,
	}
}

func (msg *MsgRepayRental) ValidateBasic() error {
	_, err := sdk.AccAddressFromBech32(msg.Creator)
	if err != nil {
		return errorsmod.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)
	}
	return nil
}
